reorg tests and add comments for whats failing and why
